modified folder structure to accomodatenew naming convention. Added ingress to helm chart
