+++ .gitremotes
@@ -1,1 +1,1 @@
-github git@github.com:dls-controls/diffcalc_API.git
+github git@github.com:DiamondLightSource/diffcalc-api.git
