+++ CONTRIBUTING.rst
@@ -7,15 +7,15 @@ filing a new one. If you have a great idea but it involves big changes, please
 file a ticket before making a pull request! We want to make sure you don't spend
 your time coding something that might not fit the scope of the project.
 
-.. _GitHub: https://github.com/dls-controls/diffcalc_API/issues
+.. _GitHub: git@github.com:DiamondLightSource/diffcalc-api.git
 
 Running the tests
 -----------------
 
 To get the source source code and run the unit tests, run::
 
-    $ git clone git://github.com/dls-controls/diffcalc_API.git
-    $ cd diffcalc_API
+    $ git clone git@github.com:DiamondLightSource/diffcalc-api.git
+    $ cd diffcalc-api
     $ pip install -e .[dev]
     $ pytest
 
@@ -102,7 +102,7 @@ To make a new release, please follow this checklist:
 - Push to internal gitlab and do a dls-release.py of the tag
 - Check and edit for clarity the autogenerated GitHub release_
 
-.. _release: https://github.com/dls-controls/diffcalc_API/releases
+.. _release: https://github.com/DiamondLightSource/diffcalc-api/releases
 
 Updating the tools
 ------------------
