+++ Dockerfile
@@ -1,7 +1,7 @@
 #Install python and copy over repo files
 FROM python:3.8
 
-ENV ENV_DIR /diffcalc_API
+ENV ENV_DIR /diffcalc_api
 WORKDIR ${ENV_DIR}
 
 COPY . ${ENV_DIR}
@@ -11,4 +11,4 @@ RUN pip install . -c requirements.txt
 
 
 #Run the API
-CMD ["uvicorn", "src.diffcalc_API.server:app", "--host", "0.0.0.0"]
+CMD ["uvicorn", "src.diffcalc_api.server:app", "--host", "0.0.0.0"]
