+++ README.rst
@@ -1,4 +1,4 @@
-diffcalc_API
+diffcalc-api
 ===========================
 
 |code_ci| |docs_ci| |coverage| |pypi_version| |license|
@@ -6,10 +6,9 @@ diffcalc_API
 An API which exposes the methods in diffcalc-core using FastAPI.
 
 ============== ==============================================================
-PyPI           ``pip install diffcalc_API``
-Source code    https://github.com/dls-controls/diffcalc_API
-Documentation  https://dls-controls.github.io/diffcalc_API
-Releases       https://github.com/dls-controls/diffcalc_API/releases
+Source code    https://github.com/DiamondLightSource/diffcalc-api
+Documentation  https://github.com/DiamondLightSource/diffcalc-api
+Releases       https://github.com/DiamondLightSource/diffcalc-api/releases
 ============== ==============================================================
 
 This is where you should put some images/code snippets that illustrate
@@ -18,31 +17,31 @@ introductory code here:
 
 .. code:: python
 
-    from diffcalc_API.hello import HelloClass
+    from diffcalc_api.hello import HelloClass
 
     hello = HelloClass("me")
     print(hello.format_greeting())
 
 Or if it is a commandline tool then you might put some example commands here::
 
-    diffcalc_API person --times=2
+    diffcalc_api person --times=2
 
 No other features yet. 
 
-.. |code_ci| image:: https://github.com/dls-controls/diffcalc_API/workflows/Code%20CI/badge.svg?branch=master
-    :target: https://github.com/dls-controls/diffcalc_API/actions?query=workflow%3A%22Code+CI%22
+.. |code_ci| image:: https://github.com/dls-controls/diffcalc_api/workflows/Code%20CI/badge.svg?branch=master
+    :target: https://github.com/dls-controls/diffcalc_api/actions?query=workflow%3A%22Code+CI%22
     :alt: Code CI
 
-.. |docs_ci| image:: https://github.com/dls-controls/diffcalc_API/workflows/Docs%20CI/badge.svg?branch=master
-    :target: https://github.com/dls-controls/diffcalc_API/actions?query=workflow%3A%22Docs+CI%22
+.. |docs_ci| image:: https://github.com/dls-controls/diffcalc_api/workflows/Docs%20CI/badge.svg?branch=master
+    :target: https://github.com/dls-controls/diffcalc_api/actions?query=workflow%3A%22Docs+CI%22
     :alt: Docs CI
 
-.. |coverage| image:: https://codecov.io/gh/dls-controls/diffcalc_API/branch/master/graph/badge.svg
-    :target: https://codecov.io/gh/dls-controls/diffcalc_API
+.. |coverage| image:: https://codecov.io/gh/dls-controls/diffcalc_api/branch/master/graph/badge.svg
+    :target: https://codecov.io/gh/dls-controls/diffcalc_api
     :alt: Test Coverage
 
-.. |pypi_version| image:: https://img.shields.io/pypi/v/diffcalc_API.svg
-    :target: https://pypi.org/project/diffcalc_API
+.. |pypi_version| image:: https://img.shields.io/pypi/v/diffcalc_api.svg
+    :target: https://pypi.org/project/diffcalc_api
     :alt: Latest PyPI version
 
 .. |license| image:: https://img.shields.io/badge/License-Apache%202.0-blue.svg
@@ -53,4 +52,4 @@ No other features yet.
     Anything below this line is used when viewing README.rst and will be replaced
     when included in index.rst
 
-See https://dls-controls.github.io/diffcalc_API for more detailed documentation.
+See https://dls-controls.github.io/diffcalc_api for more detailed documentation.
