+++ docs/explanations/test.rst
@@ -3,7 +3,7 @@ Why is something the way it is
 
 Often, reading the code will not explain *why* it is written that way. These
 explanations should be grouped together in articles here. They might include
-history of diffcalc_API, architectural decisions, or the
-real world tests that influenced the design of diffcalc_API.
+history of diffcalc_api, architectural decisions, or the
+real world tests that influenced the design of diffcalc_api.
 
 Nothing yet to explain.
