+++ docs/how-to/test.rst
@@ -1,7 +1,7 @@
 How to accomplish a task
 ========================
 
-Here you would explain how to use diffcalc_API to accomplish
+Here you would explain how to use diffcalc_api to accomplish
 a particular task. It doesn't have to be an exhaustive guide like the tutorials,
 just enough information to show someone who knows what they want to do, how to
 accomplish that task.
