+++ docs/reference/api.rst
@@ -1,26 +1,26 @@
 API
 ===
 
-.. automodule:: diffcalc_API
+.. automodule:: diffcalc_api
 
-    ``diffcalc_API``
+    ``diffcalc_api``
     -----------------------------------
 
-Internal API reference for diffcalc_API
+Internal API reference for diffcalc_api
 
 Mix verbose text with docstring and signature extraction by
 using ``autoclass`` and ``autofunction`` directives instead of
 ``automodule`` below.
 
-.. data:: diffcalc_API.__version__
+.. data:: diffcalc_api.__version__
     :type: str
 
     Version number as calculated by https://github.com/dls-controls/versiongit
 
-.. .. automodule:: diffcalc_API.hello
+.. .. automodule:: diffcalc_api.hello
 ..     :members:
 
-..     ``diffcalc_API.hello``
+..     ``diffcalc_api.hello``
 ..     -----------------------------------------
 
 No other modules yet.
