+++ helm/diffcalc-api/values.yaml
@@ -5,7 +5,7 @@
 replicaCount: 1
 
 image:
-  repository: ghcr.io/rayemelyanova/diffcalc_api
+  repository: ghcr.io/diamondlightsource/diffcalc-api
   pullPolicy: IfNotPresent
   # Overrides the image tag whose default is the chart appVersion.
   tag: latest
@@ -43,18 +43,20 @@ service:
   port: 80
 
 ingress:
-  enabled: false
+  enabled: true
   className: ""
   annotations:
     {}
     # kubernetes.io/ingress.class: nginx
     # kubernetes.io/tls-acme: "true"
   hosts:
-    - host: chart-example.local
+    - host: diffcalc.diamond.ac.uk
       paths:
         - path: /
           pathType: ImplementationSpecific
-  tls: []
+  tls: 
+    - hosts:
+      - diffcalc.diamond.ac.uk
   #  - secretName: chart-example-tls
   #    hosts:
   #      - chart-example.local
