+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = diffcalc_API
+name = diffcalc-api
 description = API for diffcalc-core python 3 package developed by Irakli Sikharulidze
-url = https://github.com/dls-controls/diffcalc_API
+url = https://github.com/DiamondLightSource/diffcalc-api
 author = Rose Yemelyanova
 author_email = rose.yemelyanova@diamond.ac.uk
 license = Apache License 2.0
@@ -52,7 +52,7 @@ where = src
 
 # Specify any package data to be included in the wheel below.
 # [options.package_data]
-# diffcalc_API =
+# diffcalc-api =
 #     subpackage/*.yaml
 
 [mypy]
@@ -74,7 +74,7 @@ extend-ignore =
 # Run pytest with all our checkers, and don't spam us with massive tracebacks on error
 addopts =
     --tb=native -vv --doctest-modules --doctest-glob="*.rst"
-    --cov=diffcalc_API --cov-report term --cov-report xml:cov.xml
+    --cov=diffcalc_api --cov-report term --cov-report xml:cov.xml
 # https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
 # filterwarnings = error #Uncomment later.
 # Doctest python code in docs, python code in src docstrings, test functions in tests
@@ -84,7 +84,7 @@ testpaths =
 [coverage:run]
 # This is covered in the versiongit test suite so exclude it here
 omit = */_version_git.py
-data_file = /tmp/diffcalc_API.coverage
+data_file = /tmp/diffcalc_api.coverage
 
 [coverage:paths]
 # Tests are run from installed location, map back to the src directory
