+++ src/diffcalc_api/errors/definitions.py
