+++ src/diffcalc_api/examples/__init__.py
@@ -1,5 +1,5 @@
 """Examples to use in endpoints for fastAPI docs, to make it easier to read."""
 
-from diffcalc_API.examples import ub
+from diffcalc_api.examples import ub
 
 __all__ = ["ub"]
