+++ src/diffcalc_api/models/__init__.py
@@ -1,12 +1,12 @@
 """Defines all pydantic models used for endpoint requests and responses.
 
 Ub and Hkl endpoints have specific pydantic models for their request bodies, defined in
-modules diffcalc_API.models.ub and diffcalc_API.models.hkl respectively.
+modules diffcalc_api.models.ub and diffcalc_api.models.hkl respectively.
 
-Module diffcalc_API.models.response defines general endpoint response models used by
+Module diffcalc_api.models.response defines general endpoint response models used by
 all routes.
 """
 
-from diffcalc_API.models import hkl, response, ub
+from diffcalc_api.models import hkl, response, ub
 
 __all__ = ["ub", "hkl", "response"]
