+++ src/diffcalc_api/routes/__init__.py
@@ -1,5 +1,5 @@
 """Defines all endpoints for the API."""
 
-from diffcalc_API.routes import constraints, hkl, ub
+from diffcalc_api.routes import constraints, hkl, ub
 
 __all__ = ["ub", "hkl", "constraints"]
