+++ src/diffcalc_api/services/__init__.py
@@ -1,5 +1,5 @@
 """Defines business logic for all endpoints, separately from the API logic."""
 
-from diffcalc_API.services import constraints, hkl, ub
+from diffcalc_api.services import constraints, hkl, ub
 
 __all__ = ["ub", "hkl", "constraints"]
