+++ src/diffcalc_api/stores/pickling.py
@@ -1,16 +1,16 @@
-"""Defines interactions with mongo persistence layer."""
+"""Defines interactions with a file system persistence layer."""
 
-from typing import Any, Dict, Optional
+import pickle
+from pathlib import Path
+from typing import Optional
 
 import numpy as np
 from diffcalc.hkl.calc import HklCalculation
 from diffcalc.hkl.constraints import Constraints
 from diffcalc.ub.calc import UBCalculation
-from motor.motor_asyncio import AsyncIOMotorCollection as Collection
-from pymongo.results import DeleteResult
 
-from diffcalc_API.database import database
-from diffcalc_API.errors.definitions import (
+from diffcalc_api.config import SAVE_PICKLES_FOLDER
+from diffcalc_api.errors.definitions import (
     ALL_RESPONSES,
     DiffcalcAPIException,
     ErrorCodesBase,
@@ -18,44 +18,45 @@ from diffcalc_API.errors.definitions import (
 
 
 class ErrorCodes(ErrorCodesBase):
-    """Persistence error codes.
-
-    This class defines all error codes which can be raised in the retrieval or storage
-    of HklCalculation objects.
-    """
+    """Codes which can be raised in the retrieval/storage of HklCalculation objects."""
 
     OVERWRITE_ERROR = 405
-    DOCUMENT_NOT_FOUND_ERROR = 404
+    FILE_NOT_FOUND_ERROR = 404
 
 
 class OverwriteError(DiffcalcAPIException):
     """Thrown if a HklCalculation object is created with a non-unique name."""
 
-    def __init__(self, name: str) -> None:
-        """Set detail and status code of the error."""
+    def __init__(self, name):
+        """Set detail and status code."""
         self.detail = (
-            f"Document already exists for crystal {name}!"
+            f"File already exists for crystal {name}!"
             f"\nEither delete via DELETE request to this URL "
             f"or change the existing properties. "
         )
         self.status_code = ErrorCodes.OVERWRITE_ERROR
 
 
-class DocumentNotFoundError(DiffcalcAPIException):
+class FileNotFoundError(DiffcalcAPIException):
     """Thrown if the store cannot retrieve a HklCalculation object."""
 
-    def __init__(self, name: str, action: str) -> None:
-        """Set detail and status code of the error."""
-        self.detail = f"Document for crystal {name} not found! Cannot {action}."
-        self.status_code = ErrorCodes.DOCUMENT_NOT_FOUND_ERROR
+    def __init__(self, name):
+        """Set detail and status code."""
+        self.detail = (
+            f"File for crystal {name} not found."
+            f"\nYou need to post to"
+            f" http://localhost:8000/{name}"
+            f" first to generate the pickled file.\n"
+        )
+        self.status_code = ErrorCodes.FILE_NOT_FOUND_ERROR
 
 
-class MongoHklCalcStore:
-    """Class to use mongo db as a persistence layer for the API."""
+class PicklingHklCalcStore:
+    """Class to use the file system as a persistence layer for the API."""
 
-    def __init__(
-        self,
-    ) -> None:
+    _root_directory: Path = Path(SAVE_PICKLES_FOLDER)
+
+    def __init__(self) -> None:
         """Set error codes that could be thrown during method excecution.
 
         Purely for documentation purposes.
@@ -71,16 +72,21 @@ class MongoHklCalcStore:
             name: the unique name to attribute to the object
             collection: the collection to store it inside.
         """
-        coll: Collection = database[collection if collection else "default"]
+        pickled_file = (
+            Path(SAVE_PICKLES_FOLDER) / (collection if collection else "default") / name
+        )
 
-        if await coll.find_one({"ubcalc.name": name}):
+        if (pickled_file).is_file():
             raise OverwriteError(name)
+
+        if not (pickled_file.parent).is_dir():
+            pickled_file.parent.mkdir()
 
         ubcalc = UBCalculation(name=name)
         constraints = Constraints()
         hkl = HklCalculation(ubcalc, constraints)
 
-        await coll.insert_one(hkl.asdict)
+        await self.save(name, hkl, collection)
 
     async def delete(self, name: str, collection: Optional[str]) -> None:
         """Delete a HklCalculation object.
@@ -89,13 +95,16 @@ class MongoHklCalcStore:
             name: the name by which to retrieve the object
             collection: the collection inside which it is stored.
         """
-        coll: Collection = database[collection if collection else "default"]
-        result: DeleteResult = await coll.delete_one({"ubcalc.name": name})
-        if result.deleted_count == 0:
-            raise DocumentNotFoundError(name, "delete")
+        pickled_file = (
+            Path(SAVE_PICKLES_FOLDER) / (collection if collection else "default") / name
+        )
+        if not pickled_file.is_file():
+            raise FileNotFoundError(name)
+
+        Path(pickled_file).unlink()
 
     async def save(
-        self, name: str, hkl: HklCalculation, collection: Optional[str]
+        self, name: str, calc: HklCalculation, collection: Optional[str]
     ) -> None:
         """Update a HklCalculation object.
 
@@ -103,8 +112,11 @@ class MongoHklCalcStore:
             name: the name by which to retrieve the object
             collection: the collection inside which it is stored.
         """
-        coll: Collection = database[collection if collection else "default"]
-        await coll.find_one_and_update({"ubcalc.name": name}, {"$set": hkl.asdict})
+        file_path = (
+            self._root_directory / (collection if collection else "default") / name
+        )
+        with open(file_path, "wb") as stream:
+            pickle.dump(obj=calc, file=stream)
 
     async def load(self, name: str, collection: Optional[str]) -> HklCalculation:
         """Load a HklCalculation object.
@@ -116,9 +128,13 @@ class MongoHklCalcStore:
         Returns:
             The HklCalculation object.
         """
-        coll: Collection = database[collection if collection else "default"]
-        hkl_json: Optional[Dict[str, Any]] = await coll.find_one({"ubcalc.name": name})
-        if not hkl_json:
-            raise DocumentNotFoundError(name, "load")
+        file_path = (
+            self._root_directory / (collection if collection else "default") / name
+        )
+        if not file_path.is_file():
+            raise FileNotFoundError(name)
 
-        return HklCalculation.fromdict(hkl_json)
+        with open(file_path, "rb") as stream:
+            hkl: HklCalculation = pickle.load(stream)
+
+        return hkl
