+++ src/diffcalc_api/stores/protocol.py
