+++ tests/test_ubcalc.py
@@ -8,9 +8,9 @@ from diffcalc.hkl.geometry import Position
 from diffcalc.ub.calc import UBCalculation
 from fastapi.testclient import TestClient
 
-from diffcalc_API.errors.ub import ErrorCodes
-from diffcalc_API.server import app
-from diffcalc_API.stores.protocol import HklCalcStore, get_store
+from diffcalc_api.errors.ub import ErrorCodes
+from diffcalc_api.server import app
+from diffcalc_api.stores.protocol import HklCalcStore, get_store
 from tests.conftest import FakeHklCalcStore
 
 dummy_hkl = HklCalculation(UBCalculation(name="dummy"), Constraints())
